requirements.txt
--index-url https://download.pytorch.org/whl/cu118
numpy
pandas
pydub
wandb
pytorch_ignite==0.4.13
torch==2.1.0
torch_summary==1.4.5
torchaudio==2.1.0
torchtext==0.16.0
